refactor(learn): extract the provider relay into relay.ts
The WP18 tool loop pushed drivers.ts over the gate's large-file rule —
the self-guardrail blocked it (dogfood catch, same extraction precedent
as page-css/page-js from render.ts). drivers.ts keeps the registry,
model listing, and routing; relay.ts owns the wire loop and its types.
No behavior change; all suites green.
